Improved Cart Component
